refactor: extract displayed to be a new hook
